feature: public Udev and UdevList API
Provides helper functions for the `Udev`, `UdevList`, and `UdevEntry`
public APIs.

Attempts to stay somewhat close to the original `libudev` C API, with
important differences when implementation details don't allow for a 1:1
mapping, e.g. `UdevList` is a `std::collections::LinkedList`, not a
collection of raw pointers.

Initial commit of providing a public API that is reasonably close to the
original `libudev`.