license: change to LGPL-v2 or later
Changes the license to LGPL-v2 or later after a re-reading of the `eudev`
license.

After discussion in
https://github.com/kevinmehall/nusb/issues/5#issuecomment-1923046730,
and reviewing the terms of LGPL, it seems required to also license
`udevrs` under LGPL.

`udevrs` is a very close translation of `eudev`. Regardless of how much
the project changes to be more idiomatic Rust, it will always have its
roots as a port of `eudev`.

IANAL, so to keep things simple let's just use the same license.