fix hwdb trie search matching
reading over libsystemd/sd-hwdb/sd-hwdb.c with general changes for
readability. Main culpret was the start,end vars searching the search
term for strlen rather than the prefix string. Not only was it the wrong
string to base search for null terminator on but also since the search
var is from user input (a Rust String), there is no null terminator
so the end value was always the length of the passed search.

https://github.com/cr8t/udev/issues/25